Move Snapshot type to own file
